requirements.txt
streamlit
python-dotenv
groq
requests
ddgs
